// src/app/models/formateur.model.ts
export interface FormateurResponse {
  id: number;
  nom: string;
  email: string;
  status: string;
  portfolio: string;
  specialite: string;
  bio: string;
  cvUrl: string;
  diplomeUrl: string;
  certificatUrl: string;
  attestationUrl: string;
  motivation: string;
  commentaireAdmin: string;
}